OPENDJ-3135: Connections to HTTP connection handler using SSL are slow (30 seconds)
ClientContext creation in HandlerAdapter injects the client's
certificate from the grizzly request's header SSL_CERTIFICATE_ATTR.
Problem is that requesting this header value is done by an asynchronous
task submitted to the grizzly's kernel thread pool (if grizzly
configured to use SameThreadIOStrategy, which is the case in DJ) which
is already busy/blocked by waiting I/O event. This asynchronous task has
to wait an IO poll time-out/event in order to be scheduled, delaying the
getHeaders() call.
Solution is to use the CERTIFICATES_ATTR instead (the one used in
grizzly http-servlet compatibility layer).